# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Museum Compass API
  version: "0.1.0"
  description: 企業博物館・歴史館評価アプリ API

servers:
  - url: /api
    description: Local / MSW mock

paths:
  /museums:
    get:
      summary: 施設一覧取得
      operationId: getMuseums
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Category"
          description: カテゴリで絞り込み
      responses:
        "200":
          description: 施設一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MuseumSummary"

  /museums/{id}:
    get:
      summary: 施設詳細取得
      operationId: getMuseumById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 施設詳細（レビュー含む）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MuseumDetail"
        "404":
          description: 施設が見つからない

  /museums/{id}/reviews:
    post:
      summary: レビュー投稿
      operationId: postReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rating
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 1000
      responses:
        "201":
          description: レビュー作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: バリデーションエラー
        "401":
          description: 未認証
        "404":
          description: 施設が見つからない
        "409":
          description: 重複レビュー

  /auth/register:
    post:
      summary: ユーザー登録
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: 登録成功
        "400":
          description: バリデーションエラー
        "409":
          description: メールアドレス重複

  /bookmarks:
    get:
      summary: 自分のブックマーク一覧取得
      operationId: getBookmarks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BookmarkStatus"
          description: ステータスで絞り込み（museumId との併用不可）
        - name: museumId
          in: query
          required: false
          schema:
            type: integer
          description: 指定施設のブックマーク1件取得（Bookmark | null で返却）
      responses:
        "200":
          description: |
            - status 指定時: BookmarkWithMuseum の配列
            - museumId 指定時: Bookmark | null（施設情報なし）
            - パラメータなし: 全ブックマーク（BookmarkWithMuseum の配列）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookmarkWithMuseum"
        "401":
          description: 未認証
    put:
      summary: ブックマーク作成・更新（upsert）
      operationId: upsertBookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - museumId
                - status
              properties:
                museumId:
                  type: integer
                status:
                  $ref: "#/components/schemas/BookmarkStatus"
                visitedAt:
                  type: string
                  format: date
                  description: 訪問日（VISITED のとき任意）
      responses:
        "200":
          description: ブックマーク作成・更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "400":
          description: バリデーションエラー
        "401":
          description: 未認証
        "404":
          description: 施設が見つからない

  /bookmarks/{museumId}:
    delete:
      summary: ブックマーク削除
      operationId: deleteBookmark
      parameters:
        - name: museumId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 削除成功
        "401":
          description: 未認証
        "404":
          description: ブックマークが見つからない

  /tags:
    get:
      summary: タグ一覧取得
      operationId: getTags
      responses:
        "200":
          description: タグ一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"

components:
  schemas:
    Category:
      type: string
      enum:
        - CORPORATE_MUSEUM
        - HISTORY_MUSEUM
        - SCIENCE_MUSEUM
        - INDUSTRIAL_HERITAGE
        - FACTORY_TOUR
        - CASTLE

    Tag:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string

    MuseumSummary:
      type: object
      required:
        - id
        - code
        - name
        - category
        - latitude
        - longitude
        - tags
        - isClosed
        - averageRating
        - reviewCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        code:
          type: string
          description: 公開用コード（cuid）
        name:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        description:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        websiteUrl:
          type: string
          format: uri
        admissionFee:
          type: integer
          nullable: true
          description: 入場料（大人料金、円単位。0=無料、null=不明）
        isClosed:
          type: boolean
          description: 閉館フラグ
        closedMessage:
          type: string
          description: 閉館理由・補足
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        averageRating:
          type: number
          format: double
          description: 平均評価 (1-5)。レビューがない場合は0
        reviewCount:
          type: integer
          description: レビュー数
        createdAt:
          type: string
          format: date-time
          description: 登録日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時

    MuseumDetail:
      allOf:
        - $ref: "#/components/schemas/MuseumSummary"
        - type: object
          required:
            - reviews
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            operatingHours:
              type: array
              items:
                $ref: "#/components/schemas/OperatingHours"

    OperatingHours:
      type: object
      required:
        - id
        - museumId
        - dayOfWeek
        - openTime
        - closeTime
        - isClosed
      properties:
        id:
          type: integer
        museumId:
          type: integer
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
          description: 曜日（0=日, 1=月, ..., 6=土）
        openTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: 開館時刻（HH:mm）
        closeTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: 閉館時刻（HH:mm）
        isClosed:
          type: boolean
          description: 当該曜日の休館フラグ
        note:
          type: string
          description: 補足（最終入館時刻等）

    BookmarkStatus:
      type: string
      enum:
        - WANT_TO_GO
        - VISITED

    Bookmark:
      type: object
      required:
        - id
        - userId
        - museumId
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        userId:
          type: string
        museumId:
          type: integer
        status:
          $ref: "#/components/schemas/BookmarkStatus"
        visitedAt:
          type: string
          format: date
          nullable: true
          description: 訪問日（VISITED のとき任意）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BookmarkWithMuseum:
      allOf:
        - $ref: "#/components/schemas/Bookmark"
        - type: object
          required:
            - museum
          properties:
            museum:
              $ref: "#/components/schemas/MuseumSummary"

    Review:
      type: object
      required:
        - id
        - rating
        - userId
        - museumId
        - userName
        - createdAt
      properties:
        id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        userId:
          type: string
        museumId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
          format: date-time